+++ src/components/TimeDisplay.vue
@@ -8,8 +8,11 @@
 
 <style scoped>
   div {
-    color:darkslategray;
-    margin-top:-50px;
-    font-size : 10rem;
+    color: #2c5f7c;
+    margin-top: -30px;
+    margin-bottom: -30px;
+    font-size: 10rem;
+    font-family: 'RocknRoll One', sans-serif;
+    text-shadow: 2px 2px 4px rgba(0,0,0,0.1);
   }
 </style>